# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Có mặt trên tất cả nền tảng kỹ thuật số hàng đầu như Google Play, iOS, website. SandexCare là ứng dụng dịch vụ hỗ trợ khách hàng giải tỏa những khó khăn trong cuộc sống qua cuộc gọi 1-1 với chuyên viên lắng nghe của SandexCare, đem lại cảm hứng và động lực trong cuộc sống - công ty hàng đầu về sức khỏe tâm lý cộng đồng tại Việt Nam.\nSandexCare sẽ luôn bên bạn mọi lúc khi bạn cần nhất và hãy đến với chúng tôi để cùng nhau lan toả năng lượng tích cực cho cuộc sống để những vấn đề của chính bạn trở nên đơn giản hơn.\n\nFormat date: \"2022-01-04\".\nFormat time_slot: \"00:00-06:00\".\nFormat booking_time: \"01:00\".",
        "title": "Tài liệu đặc tả api.sandexcare.com",
        "termsOfService": "http://sandexcare.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/master/send-event-ws": {
            "post": {
                "description": "Xử lý gửi tin nhắn đến 1 người qua websocket!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master"
                ],
                "summary": "API send message event web socket",
                "parameters": [
                    {
                        "description": "Input Send Event Web Socket",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputSendEventWebSocket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/notification": {
            "get": {
                "description": "API Get list notify of user token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseGetListNotification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "API tạo thống báo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Create Notification",
                "parameters": [
                    {
                        "description": "Input Create Notification",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputCreateNotification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/notification/read": {
            "post": {
                "description": "API Read notify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "parameters": [
                    {
                        "description": "Input read notify",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputReadNotification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.DataGetListNotification": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "read_flag": {
                    "type": "boolean"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "model.InputCreateNotification": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "string"
                },
                "receiver_role": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "model.InputReadNotification": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "read_all": {
                    "type": "boolean"
                }
            }
        },
...
        "model.InputSendEventWebSocket": {
            "type": "object",
            "properties": {
                "event": {
                    "type": "string"
                },
                "from_user_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ResponseGetListNotification": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DataGetListNotification"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.ResponseSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}